Forgot to check if #3 fix would actually build. It works now.

+++ NgFlowSample/Services/FlowUploadProcessor.cs
@@ -138,7 +138,7 @@ namespace NgFlowSample.Services
     /// </summary>
     class FileMetaData
     {
-        private static int ChunkIndex(int chunkNumber)
+        private static long ChunkIndex(long chunkNumber)
         {
             return chunkNumber - 1;
         }
